Fixed frame time measuing

+++ SanityEngine/src/stats/framerate_tracker.hpp
@@ -8,16 +8,16 @@
 enum class FramerateDisplayMode { FrameTime, FramesPerSecond, Both };
 
 struct FrametimeStats {
-    double average;
-    double minimum;
-    double maximum;
+    float average;
+    float minimum;
+    float maximum;
 };
 
 class FramerateTracker {
 public:
     explicit FramerateTracker(Uint32 max_num_samples_in);
 
-    void add_frame_time(double frame_time);
+    void add_frame_time(float frame_time);
 
     void log_framerate_stats(FramerateDisplayMode display_mode = FramerateDisplayMode::FrameTime) const;
 
@@ -26,5 +26,5 @@ public:
 private:
     Uint32 max_num_samples;
 
-    std::list<double> frame_times;
+    std::list<float> frame_times;
 };
